backend/requirements-windows.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements-windows.txt requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   elasticsearch
    #   google-genai
    #   httpx
    #   starlette
build==1.4.0
    # via pip-tools
certifi==2026.1.4
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
distro==1.9.0
    # via google-genai
dnspython==2.8.0
    # via pymongo
elastic-transport==9.2.1
    # via elasticsearch
elasticsearch==9.2.1
    # via -r requirements.in
fastapi==0.128.0
    # via -r requirements.in
filelock==3.20.2
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.12.0
    # via
    #   huggingface-hub
    #   torch
google-auth[requests]==2.47.0
    # via google-genai
google-genai==1.57.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   langgraph-sdk
    #   langsmith
    #   ollama
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.3
    # via -r requirements.in
langchain-core==1.2.6
    # via
    #   langchain
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-text-splitters==1.1.0
    # via -r requirements.in
langgraph==1.0.5
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.3.1
    # via langgraph
langsmith==0.6.1
    # via langchain-core
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.0
    # via
    #   scikit-learn
    #   scipy
    #   transformers
ollama==0.6.1
    # via -r requirements.in
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   transformers
pip-tools==7.5.2
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
pydantic-core==2.41.5
    # via pydantic
pymongo==4.16.0
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via elasticsearch
python-dotenv==1.2.1
    # via -r requirements.in
python-multipart==0.0.21
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain-core
    #   transformers
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   huggingface-hub
    #   langsmith
    #   requests-toolbelt
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.16.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   elastic-transport
    #   elasticsearch
    #   google-genai
starlette==0.50.0
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.8
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.3
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   elasticsearch
    #   fastapi
    #   google-genai
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   starlette
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   elastic-transport
    #   requests
uuid-utils==0.13.0
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.40.0
    # via -r requirements.in
websockets==15.0.1
    # via google-genai
wheel==0.45.1
    # via pip-tools
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
